Change filename and move scope data to config file
